# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: QATTH Product API
  version: 1.0.0-draft
  summary: Product v1 contract for AI-assisted career preparation and job discovery.
  description: 'Target contract for the production QATTH product. This contract intentionally

    extends beyond the current demo backend. Each operation declares

    x-implementation-status as implemented, partial, or planned.

    '
  x-contract-status: draft
  x-document-language: vi-VN
  x-contract-source: FastAPI runtime schema plus product requirement metadata
servers:
- url: http://localhost:8000
  description: Local backend
- url: https://api.qatthcv.com
  description: Production target
security:
- bearerAuth: []
...
paths:
...
  /v1/admin/model-configurations:
    get:
      tags:
      - Admin
      summary: List prompt and model configurations
      operationId: listModelConfigurations
      security:
      - bearerAuth:
        - admin:model:read
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_list_ModelConfigurationView__'
        default:
          $ref: '#/components/responses/Error'
      x-requirement-ids:
      - FR-ADMIN-002
      x-implementation-status: implemented
    post:
      tags:
      - Admin
      summary: Create a prompt or model configuration version
      operationId: createModelConfiguration
      security:
      - bearerAuth:
        - admin:model:write
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 8
          maxLength: 128
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateModelConfigurationRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_ModelConfigurationView_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/Error'
      x-requirement-ids:
      - FR-ADMIN-002
      x-implementation-status: implemented
  /v1/admin/job-sources:
    get:
      tags:
      - Admin
      summary: List job source health
      operationId: listAdminJobSources
      security:
      - bearerAuth:
        - admin:jobs:read
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: key
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Key
      - name: period_start
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Period Start
      - name: period_end
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Period End
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_list_JobSourceAdminView__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/Error'
      x-requirement-ids:
      - FR-ADMIN-004
      x-implementation-status: implemented
...
  /v1/ops/background-jobs:
    get:
      tags:
      - Operations
      summary: List background jobs
      operationId: listBackgroundJobs
      security:
      - bearerAuth:
        - ops:jobs:read
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_BackgroundJobPage_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/Error'
      x-requirement-ids:
      - FR-ADMIN-003
      - FR-OPS-002
      x-implementation-status: implemented
  /v1/ops/background-jobs/{job_id}/retry:
    parameters:
    - name: job_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    post:
      tags:
      - Operations
      summary: Retry a failed background job
      operationId: retryBackgroundJob
      security:
      - bearerAuth:
        - ops:jobs:write
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 8
          maxLength: 128
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetryBackgroundJobRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_BackgroundJobView_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/Error'
      x-requirement-ids:
      - FR-ADMIN-003
      x-implementation-status: implemented
...
  /v1/admin/job-sources/{source_id}:
    patch:
      tags:
      - Admin
      - Operations
      summary: Update Admin Job Source
      operationId: updateJobSource
      security:
      - HTTPBearer: []
      parameters:
      - name: source_id
        in: path
        required: true
        schema:
          type: string
          title: Source Id
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 8
          maxLength: 128
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateJobSourceRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_JobSourceAdminView_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-requirement-ids:
      - FR-ADMIN-004
      x-required-scopes:
      - admin:jobs:write
      x-implementation-status: implemented
  /v1/admin/model-configurations/{configuration_id}/activate:
    post:
      tags:
      - Admin
      - Operations
      summary: Activate Model Configuration
      operationId: activateModelConfiguration
      security:
      - HTTPBearer: []
      parameters:
      - name: configuration_id
        in: path
        required: true
        schema:
          type: string
          title: Configuration Id
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 8
          maxLength: 128
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivateModelConfigurationRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_ModelConfigurationView_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-requirement-ids:
      - FR-ADMIN-002
      - NFR-AI-006
      x-required-scopes:
      - admin:model:write
      x-implementation-status: implemented
  /v1/admin/moderation-cases:
    get:
      tags:
      - Admin
      - Operations
      summary: List Moderation Cases
      operationId: listModerationCases
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          default: open
          title: Status
      - name: source_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Id
      - name: period_start
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Period Start
      - name: period_end
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Period End
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_list_ModerationCaseView__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-requirement-ids:
      - FR-ADMIN-004
      x-required-scopes:
      - admin:jobs:read
      x-implementation-status: implemented
  /v1/admin/moderation-cases/{case_id}/resolve:
    post:
      tags:
      - Admin
      - Operations
      summary: Resolve Moderation Case
      operationId: resolveModerationCase
      security:
      - HTTPBearer: []
      parameters:
      - name: case_id
        in: path
        required: true
        schema:
          type: string
          title: Case Id
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 8
          maxLength: 128
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveModerationCaseRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_ModerationCaseView_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-requirement-ids:
      - FR-ADMIN-004
      x-required-scopes:
      - admin:jobs:write
      x-implementation-status: implemented
  /v1/admin/resources/{resource_type}/{resource_id}:
    get:
      tags:
      - Admin
      - Operations
      summary: Get Admin Resource
      operationId: getAdminResource
      security:
      - HTTPBearer: []
      parameters:
      - name: resource_type
        in: path
        required: true
        schema:
          type: string
          title: Resource Type
      - name: resource_id
        in: path
        required: true
        schema:
          type: string
          title: Resource Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_AdminResourceSummary_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-requirement-ids:
      - FR-ADMIN-001
      x-required-scopes:
      - admin:resources:read
      x-implementation-status: implemented
  /v1/admin/users:
    get:
      tags:
      - Admin
      - Operations
      summary: Search Admin Users
      operationId: searchAdminUsers
      security:
      - HTTPBearer: []
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 3
          maxLength: 320
          title: Q
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_list_AdminUserSummary__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-requirement-ids:
      - FR-ADMIN-001
      x-required-scopes:
      - admin:users:read
      x-implementation-status: implemented
  /v1/admin/users/{user_id}/status:
    patch:
      tags:
      - Admin
      - Operations
      summary: Update Admin User Status
      operationId: updateAdminUserStatus
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 8
          maxLength: 128
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountStatusRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_AccountStatusView_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-requirement-ids:
      - FR-AUTH-004
      - FR-ADMIN-001
      x-required-scopes:
      - admin:users:write
      x-implementation-status: implemented
...
  /v1/ops/diagnostics:
    get:
      tags:
      - Admin
      - Operations
      summary: Get Ops Diagnostics
      operationId: getOpsDiagnostics
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_OpsDiagnosticsView_'
      security:
      - HTTPBearer: []
      x-requirement-ids:
      - FR-OPS-001
      - NFR-AVL-002
      x-required-scopes:
      - ops:jobs:read
      x-implementation-status: implemented
  /v1/ops/provider-usage:
    get:
      tags:
      - Admin
      - Operations
      summary: Get Provider Usage
      operationId: getProviderUsage
      security:
      - HTTPBearer: []
      parameters:
      - name: provider
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider
      - name: purpose
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Purpose
      - name: period_start
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Period Start
      - name: period_end
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Period End
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_ProviderUsageSummaryView_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-requirement-ids:
      - FR-OPS-002
      - NFR-OBS-002
      - NFR-AI-005
      x-required-scopes:
      - ops:jobs:read
      x-implementation-status: implemented
...
  /v1/admin/model-configurations/{configuration_id}/evaluation-reports:
    get:
      tags:
      - Admin
      - Operations
      summary: List Model Evaluation Reports
      operationId: listModelEvaluationReports
      security:
      - HTTPBearer: []
      parameters:
      - name: configuration_id
        in: path
        required: true
        schema:
          type: string
          title: Configuration Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_list_ModelEvaluationReportView__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-requirement-ids:
      - FR-ADMIN-002
      - NFR-AI-002
      - NFR-AI-003
      - NFR-AI-006
      x-required-scopes:
      - admin:model:read
      x-implementation-status: implemented
    post:
      tags:
      - Admin
      - Operations
      summary: Create Model Evaluation Report
      operationId: createModelEvaluationReport
      security:
      - HTTPBearer: []
      parameters:
      - name: configuration_id
        in: path
        required: true
        schema:
          type: string
          title: Configuration Id
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 8
          maxLength: 128
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateModelEvaluationReportRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_ModelEvaluationReportView_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-requirement-ids:
      - FR-ADMIN-002
      - NFR-AI-002
      - NFR-AI-003
      - NFR-AI-006
      x-required-scopes:
      - admin:model:write
      x-implementation-status: implemented
components:
  securitySchemes:
    bearerAuth:
      type: oauth2
      description: OIDC access token. Scopes are enforced for privileged operations.
      flows:
        authorizationCode:
          authorizationUrl: https://identity.example.com/oauth2/authorize
          tokenUrl: https://identity.example.com/oauth2/token
          scopes:
            admin:model:read: Read AI configuration.
            admin:model:write: Write AI configuration.
            admin:jobs:read: Read job source operations.
            admin:billing:read: Read billing catalog and operations.
            admin:billing:write: Adjust credits.
            ops:jobs:read: Read background jobs.
            ops:jobs:write: Retry background jobs.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
      description: Provider-specific signature validated against the raw request body.
    HTTPBearer:
      type: http
      scheme: bearer
...
  responses:
    Error:
      description: Standard error response.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    CreateModelConfigurationRequest:
      properties:
        purpose:
          type: string
          maxLength: 120
          minLength: 2
          pattern: ^[a-z][a-z0-9_]{1,119}$
          title: Purpose
        version:
          type: string
          maxLength: 80
          minLength: 1
          title: Version
        provider:
          type: string
          enum:
          - openai
          - gemini
          title: Provider
        model:
          type: string
          maxLength: 160
          minLength: 1
          title: Model
        configuration:
          additionalProperties: true
          type: object
          title: Configuration
        output_schema_version:
          anyOf:
          - type: string
            maxLength: 80
          - type: 'null'
          title: Output Schema Version
      additionalProperties: false
      type: object
      required:
      - purpose
      - version
      - provider
      - model
      - configuration
      title: CreateModelConfigurationRequest
...
    APIResponse_BackgroundJobPage_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/BackgroundJobPage'
          - type: 'null'
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/Meta'
      type: object
      required:
      - data
      - meta
      title: APIResponse[BackgroundJobPage]
    APIResponse_BackgroundJobView_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/BackgroundJobView'
          - type: 'null'
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/Meta'
      type: object
      required:
      - data
      - meta
      title: APIResponse[BackgroundJobView]
...
    APIResponse_ModelConfigurationView_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/ModelConfigurationView'
          - type: 'null'
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/Meta'
      type: object
      required:
      - data
      - meta
      title: APIResponse[ModelConfigurationView]
...
    APIResponse_list_JobSourceAdminView__:
      properties:
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/JobSourceAdminView'
            type: array
          - type: 'null'
          title: Data
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/Meta'
      type: object
      required:
      - data
      - meta
      title: APIResponse[list[JobSourceAdminView]]
    APIResponse_list_ModelConfigurationView__:
      properties:
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/ModelConfigurationView'
            type: array
          - type: 'null'
          title: Data
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/Meta'
      type: object
      required:
      - data
      - meta
      title: APIResponse[list[ModelConfigurationView]]
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RetryBackgroundJobRequest:
      properties:
        reason:
          type: string
          maxLength: 500
          minLength: 3
          title: Reason
      additionalProperties: false
      type: object
      required:
      - reason
      title: RetryBackgroundJobRequest
...
    UpdateJobSourceRequest:
      properties:
        status:
          anyOf:
          - type: string
            enum:
            - active
            - degraded
            - disabled
          - type: 'null'
          title: Status
        quality_score:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Quality Score
        reason:
          type: string
          maxLength: 500
          minLength: 3
          title: Reason
      additionalProperties: false
      type: object
      required:
      - reason
      title: UpdateJobSourceRequest
...